backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.0.1
groq==0.9.0
jinja2==3.1.4
aiofiles==23.2.1
python-dotenv==1.0.1
gunicorn==22.0.0

